Tracking code is more modular
