+++ build.gradle
@@ -25,8 +25,8 @@ version = "1.8.9-2.0.0-b6"
 group= "practicalities" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "practicalities"
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 minecraft {
     version = "1.8.9-11.15.0.1694"
